requirements.txt
keyboard
Pillow
pytesseract
fugashi
unidic-lite
pywin32
